requirements.txt
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-JWT-Simple==0.0.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonify==0.5
MarkupSafe==2.0.1
PyJWT==2.3.0
pyodbc==4.0.32
PyYAML==6.0
requests==2.26.0
secure==0.1.8
SQLAlchemy==1.4.27
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==2.3.3
